Struct dashu_ratio::RBig

#[repr(transparent)]
pub struct RBig(_);

An arbitrary precision rational number.

This struct represents an rational number with arbitrarily large numerator and denominator based on UBig and IBig.

Implementations

impl RBig

pub fn to_f32_fast(&self) -> f32

Convert the rational number to a f32.

The rounding follows the default IEEE 754 behavior (rounds to nearest, ties to even).

The rounding will be correct at most of the time, but in rare cases the mantissa can be off by one bit. Use RBig::to_f32_fast for ensured correct rounding.

Examples

assert_eq!(RBig::ONE.to_f32_fast(), 1f32);

let r = RBig::from_parts(22.into(), 7u8.into());
assert_eq!(r.to_f32_fast(), 22./7.)

pub fn to_f64_fast(&self) -> f64

Convert the rational number to a f64.

The rounding follows the default IEEE 754 behavior (rounds to nearest, ties to even).

The rounding will be correct at most of the time, but in rare cases the mantissa can be off by one bit. Use RBig::to_f32_fast for ensured correct rounding.

Examples

assert_eq!(RBig::ONE.to_f64_fast(), 1f64);

let r = RBig::from_parts(22.into(), 7u8.into());
assert_eq!(r.to_f64_fast(), 22./7.)

pub fn to_f32(&self) -> Approximation<f32, Sign>

Convert the rational number to a f32 with guaranteed correct rounding.

The rounding follows the default IEEE 754 behavior (rounds to nearest, ties to even).

Because of the guaranteed rounding, it might take a long time to convert when the numerator and denominator are large. In this case RBig::to_f32_fast can be used if the correct rounding is not required.

Examples

assert_eq!(RBig::ONE.to_f32(), Exact(1f32));

let r = RBig::from_parts(22.into(), 7u8.into());
// f32 representation of 22/7 is smaller than the actual 22/7
assert_eq!(r.to_f32(), Inexact(22./7., Negative));

pub fn to_f64(&self) -> Approximation<f64, Sign>

Convert the rational number to a f64 with guaranteed correct rounding.

The rounding follows the default IEEE 754 behavior (rounds to nearest, ties to even).

Because of the guaranteed rounding, it might take a long time to convert when the numerator and denominator are large. In this case RBig::to_f32_fast can be used if the correct rounding is not required.

Examples

assert_eq!(RBig::ONE.to_f64(), Exact(1f64));

let r = RBig::from_parts(22.into(), 7u8.into());
// f64 representation of 22/7 is smaller than the actual 22/7
assert_eq!(r.to_f64(), Inexact(22./7., Negative));

pub fn to_int(&self) -> Approximation<IBig, Self>

Convert the rational number to an IBig.

The conversion rounds toward zero. It’s equivalent to RBig::trunc, but it returns the fractional part if the rational number is not an integer.

Examples

let a = RBig::from_parts(22.into(), UBig::ONE);
assert_eq!(a.to_int(), Exact(IBig::from(22)));

let b = RBig::from_parts(22.into(), 7u8.into());
assert_eq!(b.to_int(), Inexact(
    IBig::from(3), RBig::from_parts(1.into(), 7u8.into())
));

impl RBig

pub fn square(&self) -> Self

Compute the square of the number (self * self).

Examples

let a = RBig::from_parts(2.into(), 3u8.into());
let a2 = RBig::from_parts(4.into(), 9u8.into());
assert_eq!(a.square(), a2);

pub fn cubic(&self) -> Self

Compute the cubic of the number (self * self * self).

Examples

let a = RBig::from_parts(2.into(), 3u8.into());
let a3 = RBig::from_parts(8.into(), 27u8.into());
assert_eq!(a.cubic(), a3);

pub fn pow(&self, n: usize) -> Self

Raise this number to a power of n.

Examples

let a = RBig::from_parts(2.into(), 3u8.into());
let a5 = RBig::from_parts(32.into(), 243u8.into());
assert_eq!(a.pow(5), a5);

impl RBig

pub fn from_str_radix(src: &str, radix: u32) -> Result<Self, ParseError>

Convert a string in a given base to RBig.

The numerator and the denominator are separated by /. src may contain an optional + prefix. Digits 10-35 are represented by a-z or A-Z.

Examples

assert_eq!(
    RBig::from_str_radix("+7ab/-sse", 32)?,
    RBig::from_parts((-7499).into(), 29582u16.into())
);

pub fn from_str_with_radix_prefix(src: &str) -> Result<(Self, u32), ParseError>

Convert a string with optional radix prefixes to RBig, return the parsed integer and radix.

src may contain an ‘+’ or - prefix before the radix prefix.

Allowed prefixes: 0b for binary, 0o for octal, 0x for hexadecimal.

If the radix prefixes for the numerator and the denominator are not the same, then a ParseError will be returned. The radix prefix for the denominator can be omitted, and the radix for the numerator will used in the conversion.

Examples

assert_eq!(RBig::from_str_with_radix_prefix("+0o17/25")?,
    (RBig::from_parts(0o17.into(), 0o25u8.into()), 8));
assert_eq!(RBig::from_str_with_radix_prefix("-0x1f/-0x1e")?,
    (RBig::from_parts(0x1f.into(), 0x1eu8.into()), 16));

impl RBig

pub const ZERO: Self = _

RBig with value 0

pub const ONE: Self = _

RBig with value 1

pub const NEG_ONE: Self = _

RBig with value -1

pub fn from_parts(numerator: IBig, denominator: UBig) -> Self

Create a rational number from a signed numerator and an unsigned denominator

Examples

assert_eq!(RBig::from_parts(IBig::ZERO, UBig::ONE), RBig::ZERO);
assert_eq!(RBig::from_parts(IBig::ONE, UBig::ONE), RBig::ONE);
assert_eq!(RBig::from_parts(IBig::NEG_ONE, UBig::ONE), RBig::NEG_ONE);

pub fn into_parts(self) -> (IBig, UBig)

Convert the rational number into (numerator, denumerator) parts.

Examples

assert_eq!(RBig::ZERO.into_parts(), (IBig::ZERO, UBig::ONE));
assert_eq!(RBig::ONE.into_parts(), (IBig::ONE, UBig::ONE));
assert_eq!(RBig::NEG_ONE.into_parts(), (IBig::NEG_ONE, UBig::ONE));

pub fn from_parts_signed(numerator: IBig, denominator: IBig) -> Self

Create a rational number from a signed numerator and a signed denominator

Examples

assert_eq!(RBig::from_parts_signed(1.into(), 1.into()), RBig::ONE);
assert_eq!(RBig::from_parts_signed(12.into(), (-12).into()), RBig::NEG_ONE);

pub const fn from_parts_const( sign: Sign, numerator: DoubleWord, denominator: DoubleWord ) -> Self

Create a rational number in a const context

The magnitude of the numerator and the denominator is limited to a DoubleWord.

Examples

const ONE: RBig = RBig::from_parts_const(Sign::Positive, 1, 1);
assert_eq!(ONE, RBig::ONE);
const NEG_ONE: RBig = RBig::from_parts_const(Sign::Negative, 1, 1);
assert_eq!(NEG_ONE, RBig::NEG_ONE);

pub fn numerator(&self) -> &IBig

Get the numerator of the rational number

Examples

assert_eq!(RBig::ZERO.numerator(), &IBig::ZERO);
assert_eq!(RBig::ONE.numerator(), &IBig::ONE);

pub fn denominator(&self) -> &UBig

Get the denominator of the rational number

Examples

assert_eq!(RBig::ZERO.denominator(), &UBig::ONE);
assert_eq!(RBig::ONE.denominator(), &UBig::ONE);

pub fn relax(self) -> Relaxed

Convert this rational number into a Relaxed version

Examples

assert_eq!(RBig::ZERO.relax(), Relaxed::ZERO);
assert_eq!(RBig::ONE.relax(), Relaxed::ONE);

pub const fn is_zero(&self) -> bool

Check whether the number is 0

Examples

assert!(RBig::ZERO.is_zero());
assert!(!RBig::ONE.is_zero());

pub const fn is_one(&self) -> bool

Check whether the number is 1

Examples

assert!(!RBig::ZERO.is_one());
assert!(RBig::ONE.is_one());

impl RBig

pub fn split_at_point(self) -> (IBig, Self)

Split the rational number into integral and fractional parts (split at the radix point).

It’s return is equivalent to (self.trunc(), self.fract())

Examples

assert_eq!(RBig::ONE.split_at_point(), (IBig::ONE, RBig::ZERO));

let a = RBig::from_parts(22.into(), 7u8.into());
let (trunc, fract) = a.split_at_point();
assert_eq!(trunc, IBig::from(3));
assert_eq!(fract, RBig::from_parts(1.into(), 7u8.into()));

pub fn ceil(&self) -> IBig

Compute the least integer that is greater than or equal to this number.

Examples

assert_eq!(RBig::ONE.ceil(), IBig::ONE);

let a = RBig::from_parts(22.into(), 7u8.into());
assert_eq!(a.ceil(), IBig::from(4));

pub fn floor(&self) -> IBig

Compute the greatest integer that is less than or equal to this number.

Examples

assert_eq!(RBig::ONE.floor(), IBig::ONE);

let a = RBig::from_parts(22.into(), 7u8.into());
assert_eq!(a.floor(), IBig::from(3));

pub fn round(&self) -> IBig

Compute the integer that closest to this number.

It will return the one farther from zero when the number has the 1/2 as its fractional part.

Examples

assert_eq!(RBig::ONE.round(), IBig::ONE);

let a = RBig::from_parts(22.into(), 7u8.into());
assert_eq!(a.round(), IBig::from(3));

pub fn trunc(&self) -> IBig

Returns the integral part of the rational number.

It’s guaranteed that self == self.trunc() + self.fract().

Examples

assert_eq!(RBig::ONE.trunc(), IBig::ONE);

let a = RBig::from_parts(22.into(), 7u8.into());
assert_eq!(a.trunc(), IBig::from(3));

pub fn fract(&self) -> Self

Returns the fractional part of the rational number

It’s guaranteed that self == self.trunc() + self.fract().

Examples

assert_eq!(RBig::ONE.fract(), RBig::ZERO);

let a = RBig::from_parts(22.into(), 7u8.into());
assert_eq!(a.fract(), RBig::from_parts(1.into(), 7u8.into()));

impl RBig

pub const fn sign(&self) -> Sign

Get the sign of the number. Zero value has a positive sign.

Examples

assert_eq!(RBig::ZERO.sign(), Sign::Positive);
assert_eq!(RBig::ONE.sign(), Sign::Positive);
assert_eq!(RBig::NEG_ONE.sign(), Sign::Negative);

pub const fn signum(&self) -> Self

A number representing the sign of self.

RBig::ONE if the number is positive (including inf)

RBig::ZERO if the number is zero

RBig::NEG_ONE if the number is negative (including -inf)

Examples

let r = RBig::from_parts((-10).into(), 5u8.into());
assert_eq!(r.signum(), RBig::NEG_ONE);

impl RBig

pub fn simplest_from_f32(f: f32) -> Option<Self>

Find the simplest rational number in the rounding interval of the f32 number.

This method returns None when the floating point value is not representable by a rational number, such as infinities or nans.

See RBig::simplest_in for the definition of simplicity.

The rounding interval of a f32 value is an interval such that all numbers in this range will rounded to this f32 value. For example the rounding interval for 1f32 is [1. - f32::EPSILON / 2, 1. + f32::EPSILON / 2]. That is, the error of result value will be less than 1/2 ULP.

This method can be used to recover the original fraction represented as a division of f32. See the examples below.

Examples

assert_eq!(
    RBig::simplest_from_f32(1e-1).unwrap(),
    RBig::from_parts(1.into(), 10u8.into())
);
assert_eq!(
    RBig::simplest_from_f32(22./7.).unwrap(),
    RBig::from_parts(22.into(), 7u8.into())
);

pub fn simplest_from_f64(f: f64) -> Option<Self>

Find the simplest rational number in the rounding interval of the f64 number.

This method returns None when the floating point value is not representable by a rational number, such as infinities or nans.

See RBig::simplest_in for the definition of simplicity.

The rounding interval of a f64 value is an interval such that all numbers in this range will rounded to this f64 value. For example the rounding interval for 1f64 is [1. - f64::EPSILON / 2, 1. + f64::EPSILON / 2]. That is, the error of result value will be less than 1/2 ULP.

This method can be used to recover the original fraction represented as a division of f64. See the examples below.

Examples

assert_eq!(
    RBig::simplest_from_f64(1e-1).unwrap(),
    RBig::from_parts(1.into(), 10u8.into())
);
assert_eq!(
    RBig::simplest_from_f64(22./7.).unwrap(),
    RBig::from_parts(22.into(), 7u8.into())
);

pub fn simplest_in(lower: Self, upper: Self) -> Self

Find the simplest rational number in the open interval (lower, upper).

A rational n₁/d₁ is simpler than another rational number n₂/d₂ if:

d₁ < d₂ (compare denominator)

or |n₁| < |n₂| (then compare the magnitude of numerator)

or n₂ < 0 < n₁ (then compare the sign)

lower and upper will be swapped if necessary. If lower and upper are the same number, then this number will be directly returned.

Examples

let a = RBig::from_parts(1234.into(), 5678u16.into());
let b = RBig::from_parts(1235.into(), 5679u16.into());
let s = RBig::simplest_in(a, b);
// 1234/5678 < 5/23 < 1235/5679
assert_eq!(s, RBig::from_parts(5.into(), 23u8.into()));

pub fn nearest(&self, limit: &UBig) -> Approximation<Self, Sign>

Find the closest rational number to this number with a limit of the denominator.

If the denominator of this number is larger than the limit, then it returns the closest one between self.next_up() and self.next_down() to self. If the denominator of this number is already less than or equal to the limit, then Exact(self) will be returned.

The error |self - self.nearest()| will be less than 1/(2*limit), and the sign of the error self - self.nearest() will be returned if the result is not Exact.

Examples

let a: RBig = 3.141592653.try_into().unwrap();
assert_eq!(a.nearest(&10u8.into()), Inexact(
    RBig::from_parts(22.into(), 7u8.into()),
    Sign::Positive // 22/7 > 3.141592653
));

pub fn next_up(&self, limit: &UBig) -> Self

Find the closest rational number that is greater than this number and has a denominator less than limit.

It’s equivalent to finding the next element in Farey sequence of order limit. The error |self - self.next_up()| will be less than 1/limit.

Examples

let a: RBig = 3.141592653.try_into().unwrap();
assert_eq!(a.next_up(&10u8.into()), RBig::from_parts(22.into(), 7u8.into()));

pub fn next_down(&self, limit: &UBig) -> Self

Find the closest rational number that is less than this number and has a denominator less than limit.

It’s equivalent to finding the previous element in Farey sequence of order limit. The error |self - self.next_down()| will be less than 1/limit.

Examples

let a: RBig = 3.141592653.try_into().unwrap();
assert_eq!(a.next_down(&10u8.into()), RBig::from_parts(25.into(), 8u8.into()));

impl RBig

pub fn to_float<R: Round, const B: Word>( &self, precision: usize ) -> Rounded<FBig<R, B>>

Convert the rational number to a FBig with guaranteed correct rounding.

Examples

use dashu_float::{DBig, round::Rounding::*};

assert_eq!(RBig::ONE.to_float(1), Exact(DBig::ONE));
assert_eq!(RBig::from(1000).to_float(4), Exact(DBig::from(1000)));
assert_eq!(RBig::from_parts(1000.into(), 6u8.into()).to_float(4),
    Inexact(DBig::from_parts(1667.into(), -1), AddOne));