Struct dashu_ratio::Relaxed

pub struct Relaxed(/* private fields */);

An arbitrary precision rational number without strict reduction.

This struct is almost the same as RBig, except for that the numerator and the denominator are allowed to have common divisors other than a power of 2. This allows faster computation because Gcd is not required for each operation.

Since the representation is not canonicalized, Hash is not implemented for Relaxed. Please use RBig if you want to store the rational number in a hash set, or use num_order::NumHash.

Conversion from/to RBig

To convert from RBig, use RBig::relax(). To convert to RBig, use Relaxed::canonicalize().

Implementations

impl Relaxed

pub fn to_f32_fast(&self) -> f32

Convert the rational number to a f32.

See RBig::to_f32_fast for details.

pub fn to_f64_fast(&self) -> f64

Convert the rational number to a f64.

See RBig::to_f64_fast for details.

pub fn to_f32(&self) -> Approximation<f32, Sign>

Convert the rational number to a f32 with guaranteed correct rounding.

See RBig::to_f32 for details.

pub fn to_f64(&self) -> Approximation<f64, Sign>

Convert the rational number to a f64 with guaranteed correct rounding.

See RBig::to_f64 for details.

pub fn to_int(&self) -> Approximation<IBig, Self>

Convert the rational number to am IBig.

See RBig::to_int for details.

impl Relaxed

pub fn sqr(&self) -> Self

Compute the square of the number (self * self).

See RBig::sqr for details.

pub fn cubic(&self) -> Self

Compute the cubic of the number (self * self * self).

See RBig::cubic for details.

pub fn pow(&self, n: usize) -> Self

Raise this number to a power of n.

See RBig::pow for details.

impl Relaxed

pub fn from_str_radix(src: &str, radix: u32) -> Result<Self, ParseError>

Convert a string in a given base to Relaxed.

See RBig::from_str_radix for details.

pub fn from_str_with_radix_prefix(src: &str) -> Result<(Self, u32), ParseError>

Convert a string with optional radix prefixes to RBig, return the parsed integer and radix.

See RBig::from_str_with_radix_prefix for details.

impl Relaxed

pub const ZERO: Self = _

Relaxed with value 0

pub const ONE: Self = _

Relaxed with value 1

pub const NEG_ONE: Self = _

Relaxed with value -1

pub fn from_parts(numerator: IBig, denominator: UBig) -> Self

Create a rational number from a signed numerator and a signed denominator

See RBig::from_parts for details.

pub fn into_parts(self) -> (IBig, UBig)

Convert the rational number into (numerator, denumerator) parts.

See RBig::into_parts for details.

pub fn from_parts_signed(numerator: IBig, denominator: IBig) -> Self

Create a rational number from a signed numerator and a signed denominator

See RBig::from_parts_signed for details.

pub const fn from_parts_const( sign: Sign, numerator: DoubleWord, denominator: DoubleWord, ) -> Self

Create a rational number in a const context

See RBig::from_parts_const for details.

pub fn numerator(&self) -> &IBig

Get the numerator of the rational number

See RBig::numerator for details.

pub fn denominator(&self) -> &UBig

Get the denominator of the rational number

See RBig::denominator for details.

pub fn canonicalize(self) -> RBig

Convert this rational number into an RBig version

Examples

assert_eq!(Relaxed::ONE.canonicalize(), RBig::ONE);

let r = Relaxed::from_parts(10.into(), 5u8.into());
assert_eq!(r.canonicalize().numerator(), &IBig::from(2));

pub const fn is_zero(&self) -> bool

Check whether the number is 0

See RBig::is_zero for details.

pub fn is_one(&self) -> bool

Check whether the number is 1

See RBig::is_one for details.

impl Relaxed

pub fn split_at_point(self) -> (IBig, Self)

Split the rational number into integral and fractional parts (split at the radix point).

See RBig::split_at_point for details.

pub fn ceil(&self) -> IBig

Compute the smallest integer that is greater than this number.

See RBig::ceil for details.

pub fn floor(&self) -> IBig

Compute the largest integer that is less than or equal to this number.

See RBig::floor for details.

pub fn round(&self) -> IBig

Compute the integer that closest to this number.

See RBig::round for details.

pub fn trunc(&self) -> IBig

Returns the integral part of the rational number.

See RBig::trunc for details.

pub fn fract(&self) -> Self

Returns the fractional part of the rational number

See RBig::fract for details.

impl Relaxed

pub const fn sign(&self) -> Sign

Get the sign of the number. Zero value has a positive sign.

See RBig::sign for details.

pub const fn signum(&self) -> Self

A number representing the sign of self.

See RBig::signum for details.

impl Relaxed

pub fn to_float<R: Round, const B: Word>( &self, precision: usize, ) -> Rounded<FBig<R, B>>

Convert the rational number to a FBig with guaranteed correct rounding.

See RBig::to_float for details.